Split up FE files and improve view tab
